+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.5.4
+
+- Respect `WP_REDIS_SCHEME` for Cluster connections
+- Fixed issue with Predis and `SentinelReplication` connection
+- Fixed double-slash in `admin.css` URL
+
 ## 2.5.3
 
 - Added `WP_REDIS_DISABLE_GROUP_FLUSH` constant